4.23 done

+++ controllers/blogs.js
@@ -1,16 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('./../models/blog')
 const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
     const body = request.body
     const user = request.user
-    
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -40,16 +41,23 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
     const user = request.user
 
     const blog = await Blog.findById(request.params.id)
-    if (blog.user.toString() === user.id.toString()) {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+
+    if (blog) {
+        if (blog.user.toString() === user.id.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(400).json("Deleting a blog is attempted without a token or by an invalid user.")
+        }
     } else {
-        response.status(400).json("Deleting a blog is attempted without a token or by an invalid user.")
+        response.status(404).json("This blog was already deleted.")
     }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
